为了修正怪物面向方向不正确的问题，把怪的x,y，vx,vy改为了double
git-svn-id: https://viruscollector/svn/Spacecraft@241 6ae6004f-9e4d-49be-878d-5da08febfbad

+++ src/spacecraftelements/Enemy/Evasion/Basickamikaze.java
@@ -8,7 +8,7 @@ import spacecraftelements.SpecialEffect.SpecialEffect;
 public class Basickamikaze extends Enemy{
 	public double angle;
 
-	public Basickamikaze(int x,int y,double initialangle){
+	public Basickamikaze(double x,double y,double initialangle){
 		super.x = x;
 		super.y = y;
 		super.damage = 1;
@@ -38,8 +38,8 @@ public class Basickamikaze extends Enemy{
 		ratiox = Math.cos(angle*Math.PI/180);
 	    ratioy = Math.sin(angle*Math.PI/180);
 		}
-		vx = (int) (((double) v) * ratiox);
-		vy = (int) (((double) v) * ratioy);
+		vx = (v * ratiox);
+		vy =  (v * ratioy);
 		x = x + vx;
 		y = y + vy;
