+++ src/spacecraftelements/Items/S_repair.java
@@ -3,10 +3,10 @@ package spacecraftelements.Items;
 import spacecraftcore.MainGame;
 
 public class S_repair extends SpaceItem{
-	public S_repair(int x,int y)
+	public S_repair(double x,double y)
 	{
-		super.x=x;
-		super.y=y;
+		super.x=(int) x;
+		super.y=(int) y;
 		super.imageID="Images//Item//repair.png";
 		super.imagesize=25;
 	}
